+++ src/components/carousel/index.ts
@@ -1,0 +1,4 @@
+export { Carousel } from "./carousel.component";
+export { default as Slide } from "./slide";
+export type { CarouselProps } from "./carousel.component";
+export type { SlideProps } from "./slide";