+++ src/components/carousel/slide/slide.d.ts
@@ -1,9 +1,0 @@
-import * as React from "react";
-
-export interface SlideProps {
-  theme?: Record<string, unknown>;
-}
-
-declare function Slide(props: React.PropsWithChildren<SlideProps>): JSX.Element;
-
-export default Slide;